+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `VertexGroupReference.perturb` to perturb plane normals.
+- `VertexGroupReference.inject_points` to inject and overwrite points.
 
 ### Changed
 - `VertexGroupReference.bounds` &rarr; `VertexGroupReference.aabbs`.